backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.3.4
redis==5.0.1
djangorestframework-simplejwt==5.3.1
Pillow==10.1.0
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Development
black==23.12.1
flake8==7.0.0
isort==5.13.2